pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "air-quality-ml-dashboard"
version = "1.0.0"
description = "Air Quality Risk Modeling & Environmental Pattern Analysis"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24,<2",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "streamlit>=1.28",
    "pyarrow>=14.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "app*"]
